+++ CHANGELOG.md
@@ -26,3 +26,5 @@
 * *[Beta3.rc6]*
   * [FIX]: Default mobile theme added
   * [FIX]: Lightbox, no ads - media query updated to properly center ballots at large screen sizes
+  * [FIX]: Change YouTube, DailyMotion, Vimeo urls to be protocol-relative
+